requirements.txt

fastapi==0.110.0
uvicorn[standard]==0.27.1


SQLAlchemy==2.0.28
psycopg2-binary==2.9.9  


pydantic==2.6.3
pydantic-settings==2.2.1  


python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4  


starlette==0.36.3


python-dotenv==1.0.1
python-multipart==0.0.9